Put visualization functions in their own module.
I took the functions 'visualize_clusters' and 'cluster_histogram' and put them in a file called 'visualization_module.py'
