--- Code/mol_file.mol
@@ -1,16 +1,0 @@
-
-     RDKit          2D
-
-  6  4  0  0  0  0  0  0  0  0999 V2000
-    1.2990    2.5000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
-    0.0000    0.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
-    2.5981   -0.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
-    0.0000    1.7500    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
-    1.2990    0.7500    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
-    2.5981    1.7500    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
-  1  6  1  0
-  2  5  1  0
-  3  5  1  0
-  1  4  1  0
-M  END
-
