--- Code/mol_fileOut
@@ -1,2 +1,0 @@
-mol_file has assembly index: 2
-time to completion: 0
