--- Code/mol_filePathway
@@ -1,22 +1,0 @@
-{
-"file_graph":[
-{
-"Vertices": [0,1,2,3,4,5],
-"Edges": [[0,5],[0,3],[1,4],[2,4]],
-"VertexColours": ["C","C","C","C","C","C"],
-"EdgeColours": ["single","single","single","single"]
-}
-],
-"remnant":[
-{
-"Vertices": [0,3,5],
-"Edges": [[0,5],[0,3]],
-"VertexColours": ["C","C","C"],
-"EdgeColours": ["single","single"]
-}
-],
-"duplicates":[
-{"Left":[[0,5],[0,3]],"Right":[[1,4],[2,4]]}
-],
-"removed_edges":[]
-}
